+++ README.md
@@ -1,51 +1,16 @@
-# BGLK -- Bus Booking Application 
+# bus_booking_app
 
+A new Flutter project.
 
-This is a mobile application designed to streamline the process of booking a desired bus option.
+## Getting Started
 
-## Main Features
-1.  Bus Schedule and Tracking Feature
-2.  Bus Booking Feature
-3.  Bus Help and Safety Feature
-4.  Bus News and Alerts Feature
+This project is a starting point for a Flutter application.
 
-## Tech Stack
-- **Flutter**: For building the cross-platform mobile application.
-- **Dart**: The programming language used by Flutter.
-- **Firebase**: For backend services including authentication, database, and storage.
+A few resources to get you started if this is your first Flutter project:
 
+- [Lab: Write your first Flutter app](https://docs.flutter.dev/get-started/codelab)
+- [Cookbook: Useful Flutter samples](https://docs.flutter.dev/cookbook)
 
-
-## Installation
-To get started with this project, follow these steps:
-
-1. **Clone the repository**:
-    ```sh
-    git clone https://github.com/SadeeshaPerera/bus_booking_app.git
-    ```
-2. **Navigate to the project directory**:
-    ```sh
-    cd bus_booking_app
-    ```
-3. **Install dependencies**:
-    ```sh
-    flutter pub get
-    ```
-4. **Run the app**:
-    ```sh
-    flutter run
-    ```
-
-
-
-## Contributing
-Please follow these steps to contribute:
-
-1. Fork the repository.
-2. Create a new branch (`git checkout -b feature-branch`).
-3. Make your changes.
-4. Commit your changes (`git commit -m 'Add some feature'`).
-5. Push to the branch (`git push origin feature-branch`).
-6. Open a pull request.
-
-
+For help getting started with Flutter development, view the
+[online documentation](https://docs.flutter.dev/), which offers tutorials,
+samples, guidance on mobile development, and a full API reference.
