--- lib/elevator_screen.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ElevatorScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Elevator Screen'),
-      ),
-      body: Center(
-        child: Text('Welcome to the Elevator Screen!'),
-      ),
-    );
-  }
-}
