+++ lib/home.dart
@@ -1,14 +1,7 @@
-import 'package:bus_booking_app/bus_tracking_feature/search_results.dart';
-
-import 'package:bus_booking_app/bus_tracking_feature/bus_time_table_screen.dart';
-
-import 'package:bus_booking_app/elevator_screen.dart';
-import 'package:bus_booking_app/notification_panel_screen.dart';
-
+import 'package:bus_booking_app/search_results.dart';
 import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:flutter/material.dart';
-
-import 'package:bus_booking_app/AnnouncementScreen.dart'; // Import the AnnouncementScreen
+import 'package:bus_booking_app/BusTimeTableScreen.dart';
 import 'package:bus_booking_app/t-2-bus-booking-feature/pages/sample_route_page.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -52,78 +45,44 @@ class HomeScreen extends StatelessWidget {
         ],
         automaticallyImplyLeading: false,
       ),
-      body: SingleChildScrollView(
-        child: Center(
-          child: Column(
-            children: [
-              Image.asset('assets/images/Bus_App_Logo_3.png'),
-              Text(
-                'Welcome!',
-                style: Theme.of(context).textTheme.displaySmall,
-              ),
-              const SignOutButton(),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => BusTimetableScreen()),
-                  );
-                },
-                child: Text('Bus Timetable'),
-              ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => SearchResultsScreen()),
-                  );
-                },
-                child: Text('Bus Search Results'),
-              ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => ElevatorScreen()),
-                  );
-                },
-                child: Text('Go to Elevator'),
-              ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) =>
-                            AnnouncementScreen()), // Navigate to AnnouncementScreen
-                  );
-                },
-                child: Text('Missing Items'),
-              ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) =>
-                            NotificationPanelScreen()), // Navigate to Notification Panel
-                  );
-                },
-                child: Text('Go to Notifications'),
-              ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => SampleRoutePage()),
-                  );
-                },
-                child: const Text('Bus seat booking'),
-              ),
-            ],
-          ),
+      body: Center(
+        child: Column(
+          children: [
+            Image.asset('assets/images/Bus_App_Logo_3.png'),
+            Text(
+              'Welcome!',
+              style: Theme.of(context).textTheme.displaySmall,
+            ),
+            const SignOutButton(),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => BusTimetableScreen()),
+                );
+              },
+              child: Text('Bus Timetable'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => SearchResultsScreen()),
+                );
+              },
+              child: Text('Bus Search Results'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => SampleRoutePage()),
+                );
+              },
+              child: const Text('Bus seat booking'),
+            ),
+          ],
         ),
       ),
     );
