+++ linux/flutter/generated_plugins.cmake
@@ -4,7 +4,6 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   desktop_webview_auth
-  file_selector_linux
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
