+++ pubspec.yaml
@@ -41,8 +41,6 @@ dependencies:
   google_sign_in: ^6.2.1
   firebase_ui_oauth_google: ^1.3.3
   flutter_dotenv: ^5.1.0
-  firebase_storage: ^12.3.0
-  image_picker: ^0.8.7+3
   flutter_datetime_picker: ^1.5.1
   table_calendar: ^3.0.5
   cloud_firestore: ^5.4.1
@@ -71,11 +69,6 @@ flutter:
   assets:
     - assets/images/Bus_App_Logo_3.png
     - assets/images/round_dp.png
-
-    - assets/images/news_alert.png
-    - assets/images/bus_fares.png
-    - assets/images/support_desk.png
-
     - assets/images/calendar.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
