+++ windows/flutter/generated_plugins.cmake
@@ -5,10 +5,8 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   cloud_firestore
   desktop_webview_auth
-  file_selector_windows
   firebase_auth
   firebase_core
-  firebase_storage
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
